requirements.txt
Django==1.7.1
argparse==1.2.1
django-appconf==0.6
django-geoposition==0.2.2
six==1.8.0
wsgiref==0.1.2
